Refactor classes in prep for perf improvements (#8319)
* Refactor classes in prep for perf improvements
* Fix linter issues
